Essa branch adiciona a classe concreta para implementar IAdminRepository. A classe concreta possui os métodos create, findByEmail e limparTabela;
Além disso foi implementado um Seeder para popular o banco de dados com um objeto padrão para testes.

Com essa branch é esperado que o dev consiga realizar os testes de login.

Obs: O arquivo env-example foi alterado, então é necessário fazer modificações no arquivo env local.

+++ src/models/endereco.model.ts
@@ -8,7 +8,7 @@ export class Endereco {
     @Column({type: "varchar"})
     public bairro: string;
 
-    @Column({type: "long"})
+    @Column({type: "int"})
     public numero: number;
 
     @Column({type: "varchar"})
